maven build, start of AsciiDoc
