+++ .gitignore
@@ -3,3 +3,4 @@
 /tmp
 /workspace
 /classes/
+/target/
