+++ src/main/java/org/rulez/magwas/zentatools/XPathFunctions.java
